Refactored with reexport pattern
