+++ app.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
-require('dotenv').config();
+require("dotenv").config();
 
 const contactsRouter = require("./src/routes");
 const app = express();
