+++ models/contact/index.js
@@ -1,0 +1,14 @@
+const {
+  Contact,
+  addContactValidation,
+  updateContactValidation,
+  updateFavoriteValidation
+} = require("./contact");
+
+
+module.exports = {
+  Contact,
+  addContactValidation,
+  updateContactValidation,
+  updateFavoriteValidation
+};
