+++ server.js
@@ -1,4 +1,5 @@
 const mongoose = require("mongoose");
+
 const app = require("./app");
 
 const { DB_HOST, PORT = 8080 } = process.env;
