+++ src/helpers/contactHelpers/index.js
@@ -1,10 +1,10 @@
+const asyncHandler = require("./asyncHandler");
+const mongooseErrorHandler = require("./mongooseErrorHandler");
 const {
-  asyncHandler,
-  mongooseErrorHandler,
   phoneValidationPattern,
   nameValidationPattern,
   phoneValidationMessage,
-} = require("./contactHelpers");
+} = require("./requestValidationHalpers");
 
 module.exports = {
   asyncHandler,
