+++ src/routes/api/index.js
@@ -1,0 +1,3 @@
+const router = require('./contacts');
+
+module.exports = router;
