+++ src/services/contactsServices/addContact.js
@@ -1,0 +1,10 @@
+const { Contact } = require("../../../models");
+
+const addContact = async (body) => {
+  const newContact = await Contact.create(body);
+  return newContact;
+};
+
+module.exports = {
+  addContact,
+};
