+++ src/services/contactsServices/getContactById.js
@@ -1,0 +1,10 @@
+const { Contact } = require("../../../models");
+
+const getContactById = async (contactId) => {
+  const contact = await Contact.findById(contactId);
+  return contact;
+};
+
+module.exports = {
+  getContactById,
+};
