+++ src/services/contactsServices/getListContacts.js
@@ -1,0 +1,10 @@
+const { Contact } = require("../../../models");
+
+const getListContacts = async () => {
+  const contactsList = await Contact.find();
+  return contactsList;
+};
+
+module.exports = {
+  getListContacts
+};
