+++ src/services/contactsServices/removeContact.js
@@ -1,0 +1,10 @@
+const { Contact } = require("../../../models");
+
+const removeContact = async (contactId) => {
+  const contact = await Contact.findByIdAndRemove(contactId);
+  return contact;
+};
+
+module.exports = {
+  removeContact,
+};
