+++ src/services/contactsServices/updateContact.js
@@ -1,0 +1,12 @@
+const { Contact } = require("../../../models");
+
+const updateContact = async (contactId, body) => {
+  const contact = await Contact.findByIdAndUpdate(contactId, body, {
+    new: true,
+  });
+  return contact;
+};
+
+module.exports = {
+  updateContact,
+};
